feat(051-a): lift cell composition into src/bus/cells behind LayoutOptions.cell_composition (default Off)
RFC-051 Phase A. extract.rs: engine-as-generator bootstrap +
segment-crop extraction + contiguity-port derivation. compose.rs:
corridor stamping + both calibrated composers, verbatim (parity gate).
Tests become consumers; the permanent gates pin the sim-verified
geometry (EC@15: 110x22 / 461 entities) so a lift that changes
geometry fails by construction. Superseded east-feed composer +
probes dropped (findings live in the RFC-048 log). Flag plumbed
through all LayoutOptions construction sites incl. wasm-bindings;
nothing reads it until Phase B.

Suite 888/0/45, goldens 8/8 zero diffs, clippy 0, WASM builds.
