+++ crates/core/tests/e2e.rs
@@ -316,6 +316,7 @@ fn run_e2e_inner(
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
         .map_err(|e| {
@@ -1757,6 +1758,7 @@ fn tier4_advanced_circuit_7s_horizontal_stack_belt_pipe_crossing() {
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("{test_name}: layout: {e}"));
@@ -1906,6 +1908,7 @@ fn tier5_processing_unit_2s_horizontal_stack_iron_ore_pipe_bypass() {
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("{test_name}: layout: {e}"));
@@ -2016,6 +2019,7 @@ fn tier5_processing_unit_25s_horizontal_stack_pole_coverage() {
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("{test_name}: layout: {e}"));
@@ -7035,6 +7039,7 @@ fn quality_differential_ec_normal_vs_legendary() {
                 merge_tap: false,
                 stacking: 1,
                 inserter_capacity: 0,
+                cell_composition: Default::default(),
             },
         )
         .unwrap_or_else(|e| panic!("{quality:?} layout: {e}"));
@@ -7159,6 +7164,7 @@ fn quality_ec_45s_express_legendary_from_ore() {
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("layout: {e}"));
@@ -7249,6 +7255,7 @@ fn quality_differential_kovarex_self_loop_normal_vs_legendary() {
                 merge_tap: false,
                 stacking: 1,
                 inserter_capacity: 0,
+                cell_composition: Default::default(),
             },
         )
         .unwrap_or_else(|e| panic!("{quality:?} layout: {e}"));
@@ -7388,6 +7395,7 @@ fn quality_ec_45s_legendary_tree_wire_differential() {
             merge_tap: false,
             stacking: 1,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("layout: {e}"));
@@ -7466,6 +7474,7 @@ fn stacking_ec_60s_red_one_belt_headline() {
                 merge_tap: false,
                 stacking,
                 inserter_capacity: 0,
+                cell_composition: Default::default(),
             },
         )
         .unwrap_or_else(|e| panic!("S={stacking} layout: {e}"));
@@ -7580,6 +7589,7 @@ fn stacking_fanin_wall_lift_ec6_yellow_legendary() {
         merge_tap: false,
         stacking,
         inserter_capacity: 0,
+        cell_composition: Default::default(),
     };
 
     // S=1: the fan-in wall holds — 25/s cable > 15/s full yellow.
@@ -7666,6 +7676,7 @@ fn stacking_refuses_low_inserter_cap() {
             merge_tap: false,
             stacking: 2,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .expect_err("stacking=2 with max_inserter_tier=Fast must refuse");
@@ -7714,6 +7725,7 @@ fn stacking_kovarex_family_exempt_s2() {
                 merge_tap: false,
                 stacking,
                 inserter_capacity: 0,
+                cell_composition: Default::default(),
             },
         )
         .unwrap_or_else(|e| panic!("S={stacking} layout: {e}"));
@@ -7781,6 +7793,7 @@ fn stacking_ec_60s_express_legendary_s2() {
             merge_tap: false,
             stacking: 2,
             inserter_capacity: 0,
+            cell_composition: Default::default(),
         },
     )
     .unwrap_or_else(|e| panic!("layout: {e}"));
@@ -7864,6 +7877,7 @@ fn research_l7_thins_output_inserters_s4() {
                 merge_tap: false,
                 stacking: 4,
                 inserter_capacity: level,
+                cell_composition: Default::default(),
             },
         )
         .unwrap_or_else(|e| panic!("L={level} layout: {e}"));
